✨ Link with new attestation
