+++ packages/applications/document-builder/src/candidature/attestation/index.ts
@@ -1,1 +1,2 @@
 export * as AttestationSaga from './attestation.saga';
+export { generateCertificate } from './generateCertificate';
