+++ packages/applications/document-builder/src/index.ts
@@ -3,5 +3,5 @@ export {
   DonnéesDocument,
 } from './abandon/accordAbandonAvecRecandidature/buildDocument';
 export { buildDocxDocument } from './buildDocxDocument/buildDocxDocument';
-export { AttestationSaga } from './candidature/attestation';
+export { AttestationSaga, generateCertificate } from './candidature/attestation';
 export { buildTestDocument } from './buildTestDocument/index';
