+++ packages/applications/ssr/src/app/candidatures/[identifiant]/previsualiser-attestation/route.ts
@@ -1,10 +1,17 @@
-import { buildTestDocument } from '@potentiel-applications/document-builder';
+import { notFound } from 'next/navigation';
+
 import { AccèsFonctionnalitéRefuséError } from '@potentiel-domain/utilisateur';
+import { generateCertificate } from '@potentiel-applications/document-builder';
+import { DateTime } from '@potentiel-domain/common';
 
 import { withUtilisateur } from '@/utils/withUtilisateur';
+import { IdentifiantParameter } from '@/utils/identifiantParameter';
+import { decodeParameter } from '@/utils/decodeParameter';
 
-export const GET = async () =>
+export const GET = async (_: Request, { params: { identifiant } }: IdentifiantParameter) =>
   withUtilisateur(async (utilisateur) => {
+    const identifiantProjet = decodeParameter(identifiant);
+
     const rôleUtilisateur = utilisateur.role;
 
     const canPreviewAttestation = rôleUtilisateur.aLaPermission(
@@ -18,9 +25,17 @@ export const GET = async () =>
       );
     }
 
-    const content = await buildTestDocument();
+    const certificate = await generateCertificate(
+      identifiantProjet,
+      DateTime.now().formatter(),
+      utilisateur.identifiantUtilisateur.email,
+    );
 
-    return new Response(content, {
+    if (!certificate) {
+      return notFound();
+    }
+
+    return new Response(certificate, {
       headers: {
         'content-type': 'application/pdf',
         'content-disposition': 'inline',
